# _docs/openapi.yaml
openapi: 3.0.3
info:
  title: TableQueue API
  version: 0.1.0
  description: |
    Restaurant waitlist manager API.
    All UI text is in English. Formats follow Taiwan localization.
servers:
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/staff/waitlist:
    get:
      tags: [staff]
      operationId: listWaitlist
      summary: List waitlist entries
      parameters:
        - name: status
          in: query
          # Default ACTIVE lives in the referenced schema; specs.md:156 is its source.
          schema:
            $ref: '#/components/schemas/WaitlistFilterStatus'
        - name: search
          in: query
          schema:
            type: string
          description: Name or phone last 3 digits
        - name: party_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Waitlist entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistListResponse'
              example:
                items:
                  - id: 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60
                    queue_number: A001
                    full_queue_number: A-20260910-001
                    name: John Smith
                    phone_masked: 0900-***-001
                    party_size: 4
                    note: Window seat
                    status: WAITING
                    source: CUSTOMER
                    cancelled_reason: null
                    created_at: '2026-09-10T13:00:00+08:00'
                    called_at: null
                    seated_at: null
                    remaining_seconds: null
                    table_id: null
                    table_label: null
                    updated_at: '2026-09-10T13:00:00+08:00'
                total: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/staff/waitlist/{id}/call:
    post:
      tags: [staff]
      operationId: callWaitlist
      summary: Call a waitlist entry
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              maxProperties: 0
      responses:
        '200':
          description: Entry called
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistEntryResponse'
              example:
                id: 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60
                queue_number: A001
                full_queue_number: A-20260910-001
                name: John Smith
                phone_masked: 0900-***-001
                party_size: 4
                note: Window seat
                status: CALLED
                source: CUSTOMER
                cancelled_reason: null
                created_at: '2026-09-10T13:00:00+08:00'
                called_at: '2026-09-10T13:10:00+08:00'
                seated_at: null
                remaining_seconds: 600
                hold_minutes_snapshot: 10
                table_id: null
                table_label: null
                updated_at: '2026-09-10T13:10:00+08:00'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/WaitlistNotFound'
        '409':
          $ref: '#/components/responses/WaitlistInvalidStatus'
  /api/v1/staff/waitlist/{id}/seat:
    post:
      tags: [staff]
      operationId: seatWaitlist
      summary: Seat a waitlist entry
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SeatWaitlistRequest'
            example:
              table_id: 9c1b2d3e-4f5a-6b7c-8d9e-0f1a2b3c4d5e
      responses:
        '200':
          description: Entry seated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistEntryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/WaitlistNotFound'
        '409':
          description: Table not available or invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                TableNotAvailable:
                  value:
                    error:
                      code: TABLE_NOT_AVAILABLE
                      message: Table is not available
                      details: {}
                InvalidStatus:
                  value:
                    error:
                      code: WAITLIST_INVALID_STATUS
                      message: Invalid status for this action
                      details: {}
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/staff/waitlist/{id}/no-show:
    post:
      tags: [staff]
      operationId: noShowWaitlist
      summary: Mark a waitlist entry as no-show
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              maxProperties: 0
      responses:
        '200':
          description: Entry marked no-show
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistEntryResponse'
              example:
                id: 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60
                queue_number: A001
                full_queue_number: A-20260910-001
                name: John Smith
                phone_masked: 0900-***-001
                party_size: 4
                note: Window seat
                status: NO_SHOW
                source: CUSTOMER
                cancelled_reason: null
                created_at: '2026-09-10T13:00:00+08:00'
                called_at: '2026-09-10T13:10:00+08:00'
                seated_at: null
                remaining_seconds: null
                hold_minutes_snapshot: 10
                table_id: null
                table_label: null
                updated_at: '2026-09-10T13:20:00+08:00'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/WaitlistNotFound'
        '409':
          $ref: '#/components/responses/WaitlistInvalidStatus'
  /api/v1/staff/waitlist/{id}/restore:
    post:
      tags: [staff]
      operationId: restoreWaitlist
      summary: Restore a no-show entry to waiting
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              maxProperties: 0
      responses:
        '200':
          description: Entry restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistEntryResponse'
              example:
                id: 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60
                queue_number: A001
                full_queue_number: A-20260910-001
                name: John Smith
                phone_masked: 0900-***-001
                party_size: 4
                note: Window seat
                status: WAITING
                source: CUSTOMER
                cancelled_reason: null
                created_at: '2026-09-10T13:00:00+08:00'
                called_at: '2026-09-10T13:10:00+08:00'
                seated_at: null
                remaining_seconds: null
                hold_minutes_snapshot: 10
                table_id: null
                table_label: null
                updated_at: '2026-09-10T13:25:00+08:00'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/WaitlistNotFound'
        '409':
          $ref: '#/components/responses/WaitlistInvalidStatus'
  /api/v1/staff/waitlist/{id}/revert:
    post:
      tags: [staff]
      operationId: revertWaitlist
      summary: Revert a called entry back to waiting
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              maxProperties: 0
      responses:
        '200':
          description: Entry reverted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistEntryResponse'
              example:
                id: 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60
                queue_number: A001
                full_queue_number: A-20260910-001
                name: John Smith
                phone_masked: 0900-***-001
                party_size: 4
                note: Window seat
                status: WAITING
                source: CUSTOMER
                cancelled_reason: null
                created_at: '2026-09-10T13:00:00+08:00'
                called_at: null
                seated_at: null
                remaining_seconds: null
                hold_minutes_snapshot: null
                table_id: null
                table_label: null
                updated_at: '2026-09-10T13:18:00+08:00'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/WaitlistNotFound'
        '409':
          $ref: '#/components/responses/WaitlistInvalidStatus'
  /api/v1/staff/waitlist/{id}/cancel:
    post:
      tags: [staff]
      operationId: cancelWaitlistByStaff
      summary: Cancel a waitlist entry (staff)
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              maxProperties: 0
      responses:
        '200':
          description: Entry cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistEntryResponse'
              example:
                id: 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60
                queue_number: A001
                full_queue_number: A-20260910-001
                name: John Smith
                phone_masked: 0900-***-001
                party_size: 4
                note: Window seat
                status: CANCELLED
                source: CUSTOMER
                cancelled_reason: STAFF
                created_at: '2026-09-10T13:00:00+08:00'
                called_at: null
                seated_at: null
                remaining_seconds: null
                table_id: null
                table_label: null
                updated_at: '2026-09-10T13:20:00+08:00'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/WaitlistNotFound'
        '409':
          $ref: '#/components/responses/WaitlistInvalidStatus'
  /api/v1/staff/waitlist/{id}:
    put:
      tags: [staff]
      operationId: editWaitlist
      summary: Edit party size or note
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditWaitlistRequest'
            example:
              party_size: 5
              note: Needs high chair
      responses:
        '200':
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistEntryResponse'
              example:
                id: 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60
                queue_number: A001
                full_queue_number: A-20260910-001
                name: John Smith
                phone_masked: 0900-***-001
                party_size: 5
                note: Needs high chair
                status: WAITING
                source: CUSTOMER
                cancelled_reason: null
                created_at: '2026-09-10T13:00:00+08:00'
                called_at: null
                seated_at: null
                remaining_seconds: null
                table_id: null
                table_label: null
                updated_at: '2026-09-10T13:12:00+08:00'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/WaitlistNotFound'
        '409':
          $ref: '#/components/responses/WaitlistInvalidStatus'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/staff/waitlist/reorder:
    post:
      tags: [staff]
      operationId: reorderWaitlist
      summary: Reorder waitlist entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderWaitlistRequest'
            example:
              ordered_ids:
                - 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60
                - 6a8d2b0f-1b8c-5e2f-9f2b-2c3d4e5f6a71
      responses:
        '200':
          description: Reordered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaitlistListResponse'
              example:
                items:
                  - id: 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60
                    queue_number: A001
                    full_queue_number: A-20260910-001
                    name: John Smith
                    phone_masked: 0900-***-001
                    party_size: 4
                    note: Window seat
                    status: WAITING
                    source: CUSTOMER
                    cancelled_reason: null
                    created_at: '2026-09-10T13:00:00+08:00'
                    called_at: null
                    seated_at: null
                    remaining_seconds: null
                    table_id: null
                    table_label: null
                    updated_at: '2026-09-10T13:00:00+08:00'
                  - id: 6a8d2b0f-1b8c-5e2f-9f2b-2c3d4e5f6a71
                    queue_number: A002
                    full_queue_number: A-20260910-002
                    name: John Smith
                    phone_masked: 0900-***-002
                    party_size: 2
                    note: null
                    status: WAITING
                    source: STAFF
                    cancelled_reason: null
                    created_at: '2026-09-10T13:02:00+08:00'
                    called_at: null
                    seated_at: null
                    remaining_seconds: null
                    table_id: null
                    table_label: null
                    updated_at: '2026-09-10T13:02:00+08:00'
                total: 2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/staff/waitlist/close-day:
    post:
      tags: [staff]
      operationId: closeWaitlistDay
      summary: Close the day
      description: >
        Close day (specs section 4.9): every WAITING and CALLED entry becomes
        CANCELLED with cancelled_reason=CLOSED_DAY, every SEATED entry becomes
        DONE, and every active OCCUPIED table moves to CLEANING, in one
        transaction (section 14). Staff JWT required.
      responses:
        '200':
          description: Day closed - entry and table counts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloseDayResponse'
              example:
                cancelled: 2
                done: 1
                tables_cleaning: 1
        '401':
          $ref: '#/components/responses/Unauthorized'

# ============================================================
# STAFF — TABLES
# ============================================================
...
components:
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
        example: 5f7c1a9e-9a7b-4f1e-8e1a-1b2c3d4e5f60

# ============================================================
# COMPONENTS — RESPONSES
# ============================================================
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: AUTH_TOKEN_EXPIRED
              message: Session expired. Please log in again.
              details: {}
    WaitlistNotFound:
      description: Waitlist entry not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: WAITLIST_NOT_FOUND
              message: Waitlist entry not found
              details: {}
    WaitlistInvalidStatus:
      description: Invalid status for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: WAITLIST_INVALID_STATUS
              message: Invalid status for this action
              details: {}
...
  schemas:
...
    WaitlistStatus:
      type: string
      enum: [WAITING, CALLED, SEATED, DONE, NO_SHOW, CANCELLED]
...
    WaitlistFilterStatus:
      type: string
      enum: [ACTIVE, CLOSED, ALL]
      default: ACTIVE

    # -------------------- Error --------------------
...
    SeatWaitlistRequest:
      type: object
      required: [table_id]
      properties:
        table_id:
          type: string
          format: uuid
          example: 9c1b2d3e-4f5a-6b7c-8d9e-0f1a2b3c4d5e
    EditWaitlistRequest:
      type: object
      properties:
        party_size:
          type: integer
          minimum: 1
          maximum: 20
        note:
          type: string
          maxLength: 200
...
    WaitlistEntryResponse:
      type: object
      required:
        - id
        - queue_number
        - full_queue_number
        - status
        - party_size
        - created_at
      properties:
        id:
          type: string
          format: uuid
        queue_number:
          type: string
          example: A001
        full_queue_number:
          type: string
          example: A-20260910-001
        status:
          $ref: '#/components/schemas/WaitlistStatus'
        name:
          type: string
          example: John Smith
        phone_masked:
          type: string
          example: 0900-***-001
        party_size:
          type: integer
          example: 4
        note:
          type: string
          nullable: true
          example: Window seat
        source:
          $ref: '#/components/schemas/WaitlistSource'
        cancelled_reason:
          allOf:
            - $ref: '#/components/schemas/CancelledReason'
          nullable: true
        created_at:
          type: string
          format: date-time
          example: '2026-09-10T13:00:00+08:00'
        called_at:
          type: string
          format: date-time
          nullable: true
          example: '2026-09-10T13:10:00+08:00'
        seated_at:
          type: string
          format: date-time
          nullable: true
        remaining_seconds:
          type: integer
          nullable: true
          example: 600
        hold_minutes_snapshot:
          type: integer
          nullable: true
          example: 10
        table_id:
          type: string
          format: uuid
          nullable: true
        table_label:
          type: string
          nullable: true
          example: A1
        status_token:
          type: string
          nullable: true
          description: The minted opaque status credential; only returned on join
        status_url:
          type: string
          nullable: true
          description: The bare status page path; the token never rides in a query string
          example: /status/A001
        updated_at:
          type: string
          format: date-time
    WaitlistListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WaitlistEntryResponse'
        total:
          type: integer
          example: 1
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT